Overhaul of architecture
