+++ sanic/response/__init__.py
@@ -1,7 +1,6 @@
 from .convenience import (
     empty,
     file,
-    file_browser,
     file_stream,
     html,
     json,
@@ -33,6 +32,5 @@ __all__ = (
     "file",
     "redirect",
     "file_stream",
-    "file_browser",
     "json_dumps",
 )
